Changed stuff.
Using just the mjpg streamer on the smart dashboard causes frequent
framedrops which causes it to frequently draw a pink screen with the words
no connection. The smart dashboard plugin was altered to hold the previous
frame when connection drops. Some minor changes include, changes in aestetics
from the original forked repository, and the default ip for the camera was
set to raspberrypi.local with port 1180